+++ addons/mail/views/res_users_views.xml
@@ -16,18 +16,6 @@
             </field>
         </record>
 
-        <!-- Force update of groups' user view.
-             When updating (not new install), as `share` module is not loaded yet, shared groups
-             are taken into account when getting the group list, resulting to a different
-             organisation of the `Sharing` category (booleans instead of a selection) and thus,
-             different virtual field names.
-             When applying then inheritance of the following view, the check of field names
-             failed due to these virtual field name missmatch.
-             This is only for databases updated between a6eac0f08ee9100d1a44a4068d4f580b1c7c4b2c
-             and the commit introducing this line.
-        -->
-        <function model="res.groups" name="update_user_groups_view" />
-
         <!-- Update user form !-->
         <record id="view_users_form_mail" model="ir.ui.view">
             <field name="name">res.users.form.mail</field>
