+++ addons/share/__init__.py
@@ -2,5 +2,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import ir_model
-import res_users
 import wizard
